resolve pickling error
